cupsFinalBuild.js
I think this will actually be the final build of this tiny game that took way to long. ha. changed functions to arrow notation. added a play again loop. functionFun is now actually kind of fun and actually kind of functions. And not too ugly for a tiny baby coder.